Added Spring Security Config & refactoring
Added SSC so we can figure out whether users are logged in or not.

Added log4j so we can customise the logging.

Added some extra to readme

security.xml will need changing as the current redirects only send between /home and /signup. Should be redirecting to login page

Signup page now actually creates a user in the database

Disabled CSRF checks for the time being

Fixed home and signup links on navar

Added redirect for "/" to "/home"

Fixed the typo in the hibernate config preventing access to database