Second drop for safe keeping
It is now possible to search for listings and requests by location.
On page load for either screen, the user's location is requested. If they don't want to provide it then we default to London. The system then populates ALL the requests/services across the UK.

If a user wants to search in a particular area then that will narrow down the results and reduce the lists below to only the relevant data.

Still to do:
- Collection of location data from listings
- Make it work with groups when that is done
-  Lots of tidy up
-  Pretty much all jUnits are missing...